rfac: move showcase state logic in viewmodel
