+++ lib/viewmodels/ib/ib_landing_viewmodel.dart
@@ -3,7 +3,9 @@ import 'package:mobile_app/enums/view_state.dart';
 import 'package:mobile_app/locator.dart';
 import 'package:mobile_app/models/failure_model.dart';
 import 'package:mobile_app/models/ib/ib_chapter.dart';
+import 'package:mobile_app/models/ib/ib_showcase.dart';
 import 'package:mobile_app/services/ib_engine_service.dart';
+import 'package:mobile_app/services/local_storage_service.dart';
 import 'package:mobile_app/viewmodels/base_viewmodel.dart';
 
 class IbLandingViewModel extends BaseModel {
@@ -11,6 +13,8 @@ class IbLandingViewModel extends BaseModel {
   String IB_FETCH_CHAPTERS = 'ib_fetch_chapters';
 
   final IbEngineService _ibEngineService = locator<IbEngineService>();
+  final LocalStorageService _localStorageService =
+      locator<LocalStorageService>();
 
   // Global Keys
   final GlobalKey _toc = GlobalKey(debugLabel: 'toc');
@@ -29,6 +33,52 @@ class IbLandingViewModel extends BaseModel {
   List<IbChapter> _chapters = [];
 
   List<IbChapter> get chapters => _chapters;
+
+  // ShowCaseState stores the information of whether the button which is to be
+  // showcased are clicked or not
+  IBShowCase _showCaseState;
+
+  IBShowCase get showCaseState => _showCaseState;
+
+  set showCaseState(IBShowCase updatedState) {
+    _showCaseState = updatedState;
+    notifyListeners();
+  }
+
+  void onShowCased(String key) {
+    switch (key) {
+      case 'next':
+        if (!_showCaseState.nextButton) {
+          _showCaseState = _showCaseState.copyWith(nextButton: true);
+        }
+        break;
+      case 'prev':
+        if (!_showCaseState.prevButton) {
+          _showCaseState = _showCaseState.copyWith(prevButton: true);
+        }
+        break;
+      case 'toc':
+        if (!_showCaseState.tocButton) {
+          _showCaseState = _showCaseState.copyWith(tocButton: true);
+        }
+        break;
+      case 'drawer':
+        if (!_showCaseState.drawerButton) {
+          _showCaseState = _showCaseState.copyWith(drawerButton: true);
+        }
+        break;
+    }
+    notifyListeners();
+  }
+
+  void saveShowcaseState() {
+    _localStorageService.setShowcaseState = _showCaseState.toString();
+  }
+
+  void init() {
+    _showCaseState = IBShowCase.fromJson(_localStorageService.getShowcaseState);
+    fetchChapters();
+  }
 
   Future fetchChapters() async {
     try {
